version: 0.1 -> 0.2 (#11)
* added nix shell support

enter with nix-shell command

* adjusted Core.java layout

from template style to my own style

* added code style to README.md

* added test broadcastMessage

* updated spigot API version

* test broadcastMessage: onEnable -> onLoad

line executed before the server has a chance to load the plugin so was not showing up

* Accessing the plugin with an in-game GUI (#4)

* branch distinction for in-game testing

* updated code style in README.md

* testing for GUI class

* concept for gui templating w/ JSON

not yet fully implemented

* improved javadocs for GUILoader and GUI

* removed delombok

* added GUI instance initialiser

used to create the screen (Inventory) to be modified/controlled whenever a new GUI is created.

* updated pom.xml plugin/deps versions

* added titles to GUI

* removed getScreen() note

it would be a lot of work to collect and piece together the properties into JSON, and it will not have any use for a while.

* added JSON parse for GUI template (with title field)

* added gui template demo

* added templateString param javadoc desc to GUILoader.parse()

* moved load() parse effort into try-catch

* added GUI tests (templates)

* changed default gui size: 0 -> 9

* added size option for GUI templates

* added jackson-core dependency

* enabled dependency bundling

* set default value for GUI title class var

* fixed incorrect assertion params in GUI tests

the expected was the actual, and the actual was the expected.

* fixed post-build availability of JSON template resources

before the server would look for the full path as it is in the source code and it wouldn't be able to find it.

* removed unused GUITest imports

* added branding assets (description, colour palette, logo, banner)

* concept for parsing GUI template slots

* removed GUI open logging

* added javadoc param for GUILoader parseTitle

* added javadoc for GUILoader parseSize

* added api-version to plugin.yml

this was missing, it enables spigot to know what features to have or not have.

* added slots to GUI/parsing

This means that the slots array part of the GUI template now is partially working. It only is using the default item but the slots.slot (position in the GUI) key is working

* open test/demo GUI for all

would make it easier to people to test it on their own if my username is not hardcoded in.

* fixed missing classes for GUISlot

* filled behind border for simple dev background layer

* added slots item to GUI/parsing

meaning now that if the item is valid it will show. When invalid (such as WOOL in the demo.json not being close enough to a real Material), it will show as the default item.

* added slots label to GUI/parsing

now when the label exists it will be put on the slot, if it doesn't it will default to a string with a single whitespace

* added in-game error for invalid slots item

if the material cannot be found, it will be a red box with error pre-appended to the label

* fixed leading whitespace in non-error labels

previously it would **check if it's empty and there is an error. If not, then it would pre-append whitespace.** Now the behaviour is for it to **only pre-append a whitespace if there is an error**.

* slots label no longer italic by default

* updated the folders table in README

* added 'Access to the plugin' to README

* added usage example to GUI class javadoc

* expanded on 'way to think about the program' in README

* added command to access the plugin as GUI

currently just opens the 'empty' gui template

* removed log of raw GUI template before parsing

* removed testing GUI showing onLoad

* removed test broadcast showing onLoad

* removed unused onLoad hook in Core

* removed unused imports from Core

* Removed 'access to the plugin' notes in README

* 'Create Quest' GUI option (w/ reusable screen changing) (#5)

* added main gui screen template

* disallow taking inventory slot items from GUI

* Quest.display() template: 'empty' -> 'main'

* updated docs for Chat class

* updated docs for Commandplayerquest class

* updated docs for Quest class

* updated docs for GUIUtils class

* gui slot list type: ArrayList -> HashMap

this is so it doesn't need to be iterated over to see what slot it's in

* improved GUISlot class and gui listener

the slot class has upgraded methods and the gui click listener is more precise

* more accurate comment on registering gui events

* added planning for GUI functionality (functions)

with consideration for quest stage functionality (actions)

* less gramatically ambiguous 'Dependencies' heading

* removed empty row from 'Meta Actions' table

* added create quest to main GUI

Closes #2

* Exit GUI Item (#10)

* GUI function to close the PQ screen. (#9)

* Started creation of an exit within the GUI

* fixed missing package line

* Made it so closescreen actually closes the screen...

* added class variable comments to Core

* added more comments to GUI class

* added class variable comments to GUIListener class

* added class variable comments to GUILoader class

* added class variable comments to GUISlot class

* added comments to CloseScreen class

* added class variable comments to GUIFunction class

---------

Co-authored-by: utsu-pinku <104600733+utsu-pinku@users.noreply.github.com>

* version: 0.1 -> 0.2

---------

Co-authored-by: utsu-pinku <104600733+utsu-pinku@users.noreply.github.com>