added rough architecture diagram
shows: how to separate responsibility better, how the GUI is a loop between it's builder and it's client, and a pseudocode brief overview of most parts of the plugin. This shouldn't be relied on solely to understand the plugin, but it's an okay entry point (some of the details are not valid).