Adding actions to quest stages (#19)
* added create new action button

accessing an action shows a default gui

* forming quest action objects in sequence by string ID

LinkedHashMap is like a HashMap but explicitly tracks sequence

changes formation of json object from:
```json
actions: {
  {
    "title": "action_0"
  }
}
```

to:
```json
actions: {
  "action_0": {
    "title": "action_0"
  }
}
```

* added title + back button, to quest-action-editor

* fixed serialised action object 'title' -> 'name'

this better reflects the quest object example in the docs/readme.

from:
```json
"action_0": {
  "title": "action_0"
}
```

to:
```json
"action_0": {
  "name": "action_0"
}
```

the relevancy of this field is just for the user to be able to pick
their own name for an action that is meaningful to them. While enabling
the system to always have action_[number] to use as the ID regardless of
the name. The name at the moment is set as the same as the ID as there
is no name changing mechanism on the actions yet.