added QuestClient.showFX() and realtime quest updating
Basically meaning that NPCs should (untested) now have a particle effect when they have a quest interaction available (in technical terms, when they are the foremost NPC in the entry action of the entry quest of a stage). An idea for persistence is to mutate the Quest object for user data so that their entry stage is the one which is at their progress (but that might involve storing unecessary data; unless there is a reason why info like the NPC would need to be duplicated like different NPC states for different stages of the quest?).