implemented NPC action option
this means the NPC can be changed/selected from the action editor screen. It also means the groundwork for doing it for other action options is prepared. It's also separate from the quest option ENUM; the GUI client (in this case Dynamicactioneditor) is responsible for implementing the setting based on the enum, see Dynamicactioneditor.putOptionSlots(). When there is a chat client you would expect a command to do the same thing, as to say something like Commandactioneditor would implement which setters for which action option enum.