Assigning the quest to a block (#31)
* added 'npc types' GUI w/ unimplemented 'SelectBlock' meta function

* added NPCType.BlockNPC and fixed json serialisation for quest template

* added selecting NPC block (inventory and hitting block)

* fixed NPCType mislabelling/bad variables

* added GUI reflections/style around assigning NPC to Block

* implemented exiting SelectBlock function

* add assigning NPC a type as requirement

* add npc types back button

* added 'Place NPC' GUI button

* changed GUIUtils -> MaterialUtils

more appropriate utility class for the methods inside

* added support for setting NPC block from chat

* fixed SelectBlock blocklist data

* implemented SelectBlock denylist

* added basic support for setting Block NPC locations

* added temporary inventory clear for player selecting block location

* added clarification: Place NPC -> Place NPC Block

* added materialising the NPC into the world on quest save

* added remove method for NPCType

* fixed inventory click bugging inventoryview state

* fixed SelectBlock 'exit' to cancel

* fixed SelectBlock error when PlayerInteractEvent not a block

like when interacting with the AIR, which is not a block

* removed unecessary ActionType class logic and added abstract putOptionSlots method

* added the test-server dir to .gitignore

* fixed pom.xml plugin order for build process

would previously copy unshaded jar out instead of the shaded one (with dependencies bundled)

* reorganised builder/quest directory

* fixed jumbled stage actions map

* fixed bad architecture for QuestAction

now has the granularity to add to it's QuestStage with .submit()

* moved location object: utility/object/Location -> quest/data/LocationData

far better for organisation

* removed SelectBlock onSelect scheduler for setting response

causes issues when setResponse still finishes before other operations can fully finish

* SelectBlock ignore if AIR selected

* changed SelectBlock onSelect listener work to run on main thread

* added getValue method to KeyHandler

* fixed missing template value ('Quest' for the quest title)

* added test script to .gitignore

* added vscode config to .gitignore

* moved all TODO comments to top of files

this is a better pattern, especially for long-term or mid-term TODO notes

* moved KeyHandler: utility -> utility/singleton

* added empty QuestRegistry singleton

* associated QuestStage key with QuestStage.getTitle()

* stopped using getTitle value parsing for action editor GUI title

* changed quest template serialise target: quest builder -> quest product

* added methods to QuestRegistry: add, submit, remove, replace

* removed GUISlot dependency from GUIFunction classes

* added retrieving action object in Action Editor GUI

* converted static quest editor GUI to dynamic: create-quest.json -> Dynamicquesteditor.java

* added basic interaction with registered NPCs

'registered NPCs' refers to NPCs that have been submitted to the BlockListener via the PlayerQuests class. This is triggered by submitting the Quest product the NPCs belong to, to the QuestRegistry.

Basically: when quest registered (QuestRegistry), npcs are added to the world (PlayerQuests -> BlockListener).

* added giving Players QuestClient & registering QuestClient in QuestRegistry

* added QuestClient.showFX() and realtime quest updating

Basically meaning that NPCs should (untested) now have a particle effect when they have a quest interaction available (in technical terms, when they are the foremost NPC in the entry action of the entry quest of a stage). An idea for persistence is to mutate the Quest object for user data so that their entry stage is the one which is at their progress (but that might involve storing unecessary data; unless there is a reason why info like the NPC would need to be duplicated like different NPC states for different stages of the quest?).

* fixed identifying if action is stage entry point in Dynamicqueststage

* register players as questers on reload

the return could be removed to make it also do the same for on startup but if that was necessary it would be a bug either in our onJoin or in spigots join event.

* added unregistering a BlockNPC when destroyed

* removed TODO for unregistering a BlockNPC when destroyed

* removed unused imports from ServerListener

* added comment for ServerLoadEvent import in ServerListener

* added comment for QuestRegistry import in ServerListener

* remove the quest from QuestRegistry when NPC block broken

* update QuestClient when quest removed from QuestRegistry

* fixed instantiating npc-related quest action types

* added outline for quest action options

* added registering the ServerListener in PlayerQuests

* moved putOptionSlots to Dynamicactioneditor

This is more suitable as the GUI for editing actions should be responsible for determining the GUI logic for changing that action

* implemented NPC action option

this means the NPC can be changed/selected from the action editor screen. It also means the groundwork for doing it for other action options is prepared. It's also separate from the quest option ENUM; the GUI client (in this case Dynamicactioneditor) is responsible for implementing the setting based on the enum, see Dynamicactioneditor.putOptionSlots(). When there is a chat client you would expect a command to do the same thing, as to say something like Commandactioneditor would implement which setters for which action option enum.

* fixed NPC block particle

* removing particles from QuestClient with npcParticles map

* moved filling QuestClient helper maps: addQuest() -> update()

* update QuestClient when removing a quest from it

* added setting demo NPC dialogue to Speak action

* added formatting to Speak action demo dialogue

* added SelectBlock method denylist

with: SelectMethod.HIT, SelectMethod.CHAT, SelectMethod.SELECT

* added -1 block penalty when an NPC block is selected

* removing NPC from quest when removed from world

* corrected serialisation of quest data

* renamed questnpc -> npceditor

* added delete NPC button to npceditor

* removed test default npc from QuestBuilder

* added removing quest actions

* added back button to 'NPC Selector'

* removed allowing incomplete NPC states

* added setting own dialogue for Speak action