[v0.4] Options for editing quest stage and action, including sequences/connections (#46)
* updated 'How To Get Functionality' explainers/quest action plan

* added early quest expressions/templates

* removed duplicated 'How It All Works' section

* added re-arrange buttons for actions/stages

* added ARRANGE GUI mode

currently this is unrestricted, it just lets anything be moved around anywhere in the chest inventory and in status*player inventory**git status

* disabled menu items put out of GUI (in ARRANGE mode)

* fixed taking items from GUI

fixed methods incl: throwing out of inventory, using the dragging mechanism to bypass click block (and finally) holding the block and cancelling the inventory forcing it to default to the Player inventory instead of being destroyed. These issues were fixed respectively, except the last which has a safety net sort of method, where if the GUI is closed finally, it will loop through slots and remove any items tagged with GUI="true"

* re-toggling the re-arrange button

back into the click gui mode, from arrange gui mode

* removed re-arrange buttons for stages/actions

* added 'Change Sequence' button to quest stage editor

* added a test/proof connectioneditor

* added unimplemented connectioneditor buttons

* added empty 'select connection' menu

* added 'Back' button to 'select connection' menu

* added placeholder stages for 'select connection'

* implement stages recursion in 'select connection'

* add label for 'select connection' stage buttons

* add action sub-menu outline to 'select connection'

* added back button to action sub-menu

* add actions recursion in 'select connection' sub-menu

* added setting connection data from 'connection editor'

* removed ConnectionsData debug message

* fix null getType in actiontypes/actioneditor

* added changing entry point for Quest

* fixed action in 'quest editor'

* improved 'quest stage' new action button

* fixed null diaryConnections on NPC interact

* NPC Types: Block -> (renamed to) BlockNPC

This is a breaking change sadly, but it was hell bent on not interpreting the type data under a different simple name

* added connectioneditor to actioneditor

* indicate shared/global quests in myquests

* added heading to theirquests menus

* load up questbuilder npcs with setID()

* fixed UX: some back buttons being labelled 'Exit'

* fixed unable to 'exit' in select npc block mode

* beans-tester-bonus NPCs on the same y axis

* subpar fix: changing NPC locations

* clone quest from JSON rather than former QuestBuilder class

* removed unused quest expressions

* set recovered quests to default to not shown (untoggled)

* dividers for 'queststages' to dynamically cover first two columns

* broken: added 'add stage' button

no support for multi-stage in quest diaries

* support for multi-stage in quest diaries

* add description field for GUI buttons

* Delete button on stage editor

Includes confirmation, with checks to see if possible or if stage is depended on. Shows a gray button w/ 'Cannot Delete'. Also has  a description of why, which should maybe be split on every nth whitespace.

* add a stage_0 if it doesn't already exist

* QuestBuilder.getStages() as LinkedList

* malformed quest handling improvements

still might error out to console when the error comes from parsing the JSON in Quest.fromTemplateString()

* add 'unreadable quests' indicator w/ added GUISlot.setCount

* fixed regression with cloning quests

* fixed assigning NPC block, possible race condition

usually stacks would interrupt the Listener unregistering slash cause the current re-opened GUI to dispose itself too early leading to a dead/broken GUI.

* replaced main GUI: static -> dynamic

* spring cleaning + update compiler to Java 22

* set instance on click for myquests list

* theoretical fix: GUI breaking with multiple players

* add instantiating builders map

* do not handle onClick at all for GUI if not isGUIInventory

* fix NPC interactions breaking with multiple players

multiple players interacting; or more specifically multiple diary entries with the same quest. The issue was multiple records were being obtained when trying to get the players progress in the quest from the quest diaries (this was because it was just searching by quest name, not by quest name and diary ID)

* remove static GUI (GUI template) infrastructure

removes all the JSON GUI template parsing mess

* fixed NPC 'speak' action: dialogue only showing after set

* extended NPC dialogue action, dialogue preview length

* fix empty dialogue breaking NPC speaking

* added quest action requirements + 'Speak' NPC requirement

* added NPC LocationData to validity check

* theoretical fix: add player to clearChat calls

fixes #50

* fix clearChat failing when used before function 'setUp'

* added support for more data in BlockNPC.value (half-implemented)

addresses, and basically fixes #46. But sophisticated data like custom player heads doesn't work due to a data bottleneck in NPC block selection/placement

* small fixes to quest template loading

* added catch-and-repair for missing data in myquests GUI

* fixed quest toggling removing quest from db

* added better message sending

* fixed queststage GUI behaviour when working with actions

* theoretical fix: jank when going to previous GUI from npctypes

* change some user-facing messages: ERROR -> WARN

* fixed untoggled quests adding to the QuestClient

* added PlayerQuests tutorial

* removed PLUGIN-CUTE.md template

* add the unsupported item actions to docs

* clarify 'Speak' action is not to the world