fixed cancelling chat when chat is a denied select method in SelectBlock
to explain this: SelectBlock has select by Chat, select by Hit and select by Inventory. Those aren't the exact things, check SelectMethod if you want to see the actual ones. But basically, the chat listener didn't know it should allow chat to not be cancelled if it's not an allowed method of selecting a block.