remove double-up: adding quests to each players 'Quest Diary'
Fixes an issue where if you join super early or reload a little unlucky, it will put a Quest into your diary based on the start of the quest (even if you have progress). This change works because QuestDiary's load the quests into themselves (when they are ready, since conveniently QuestRegistry is a neatly accessible singleton).