fix Dynamicitemlist adding item after deleting one
kinda weird solution, the Back button won't go back on it's own. It requires onFinish to be set to whatever the next GUI screen is. This was basically because i needed to set onFinish for the screen it was going to, and when the UpdateScreen was happening in the item editor it had no onFinish so it was getting lost. Could potentially steal the Consumer from the Director state, but really the whole onFinish vanishing is bung, maybe UpdateScreen needs to preserve the screen onFinish along with the screen name so you can go to a million screens and eventually you'll get back to the origin and the behaviour won't have disappeared.