Add reward item action and quest inventory (#104)
* fix next action starting before game code finishes

* SelectBlock -> SelectMaterial

can add a boolean to restrict to only blocks

* add RewardItem action

* fix overuse of mutable lists

* add 'Quest Inventory' w/ restocking

* add 'Out of Stock' indication for required items

* add 'Not Enough Stock' indicator if 'requiredInventory' unmet

* remove unused 'player' variable

* fix null inventory var in getInventory

* use QuestRegistry+Database for Quest Inventories

* fix Dynamicitemlist adding item after deleting one

kinda weird solution, the Back button won't go back on it's own. It requires onFinish to be set to whatever the next GUI screen is. This was basically because i needed to set onFinish for the screen it was going to, and when the UpdateScreen was happening in the item editor it had no onFinish so it was getting lost. Could potentially steal the Consumer from the Director state, but really the whole onFinish vanishing is bung, maybe UpdateScreen needs to preserve the screen onFinish along with the screen name so you can go to a million screens and eventually you'll get back to the origin and the behaviour won't have disappeared.

* fix unable to sort inventory that doesn't exist

* fix realtime responsiveness of quest inventory

* fix incorrect variable in QuestRegistry.setInventory

* add javadoc for getPredictiveInventory

* add TakeItem contributing to 'Quest Inventory'

* updated pq guide: action support list

RequestItem and RewardItem is supported now

* fix GatherItem incorrect initOptions body

* never return null from QuestRegistry.getInventory

* force into survival when using /playerquests

* fix QuestBuilder.build always overwriting current

* move 'Quest Inventory' to 'myquest' screen (before editing)