remove nuisance anti-flying barrier
placing a barrier above the block npc so no one can stand on it is a good tactic, but i don't want people to be able to abuse basically being able to spawn barriers. Also managing them when the plugin breaks isn't the most feasible. So the flying issue either needs to be fixed some other way or not addressed here (instead the server understanding a player has a ghost block under them that it administered? i can't know for sure if that's possible for them).