fixed stage 'next actions' duplicating
fixed #115 - when asking the questclient to forcefully close any ongoing actions connected to a deleted/replaced quest it would soft-stop them and they'd just proceed to the next. Kind of more like a skip function than a stop. So i added a halt to it so we can choose for it to stop for good.